+++ astropy/coordinates/transformations.py
@@ -660,7 +660,7 @@ class CompositeStaticMatrixTransform(StaticMatrixTransform):
             for m in self.matricies[1:]:
                 matrix = np.dot(np.asarray(matrix), m)
 
-        super(CompositeStaticMatrixTransform, self).__init__(self, fromsys,
+        super(CompositeStaticMatrixTransform, self).__init__(fromsys,
             tosys, matrix, priority)